refactor: implement useTimezoneLabelsTop
